requirements.txt
torch
streamlit
torchvision
streamlit
Pillow
opencv-python
numpy
